CodeTransformer-deployment/requirements.txt
streamlit
google-cloud-aiplatform
kubernetes
pyyaml
openai
black
fastapi
uvicorn